+++ src/routers/userRouter.js
@@ -1,0 +1,9 @@
+import express from "express";
+
+const userRouter = express.Router();
+
+const handleEditUser = (req, res) => res.send("Edit User");
+
+userRouter.get("/edit", handleEditUser);
+
+export default userRouter;
