+++ src/routers/videoRouter.js
@@ -1,0 +1,9 @@
+import express from "express";
+
+const videoRouter = express.Router();
+
+const handleWatchVideo = (req, res) => res.send("Watch Video");
+
+videoRouter.get("/watch", handleWatchVideo);
+
+export default videoRouter;
