+++ src/server.js
@@ -1,29 +1,14 @@
 import express from "express";
 import morgan from "morgan";
+import globalRouter from "./routers/globalRouter";
+import userRouter from "./routers/userRouter";
+import videoRouter from "./routers/videoRouter";
 
 const PORT = 4000;
 
 const app = express();
 const logger = morgan("dev");
 app.use(logger);
-
-const globalRouter = express.Router();
-
-const handleHome = (req, res) => res.send("Home");
-
-globalRouter.get("/", handleHome);
-
-const userRouter = express.Router();
-
-const handleEditUser = (req, res) => res.send("Edit User");
-
-userRouter.get("/edit", handleEditUser);
-
-const videoRouter = express.Router();
-
-const handleWatchVideo = (req, res) => res.send("Watch Video");
-
-videoRouter.get("/watch", handleWatchVideo);
 
 app.use("/", globalRouter);
 app.use("/videos", videoRouter);
